py_file_requirements.txt
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
TA-Lib==0.4.24
python-binance==1.0.16
python-dotenv==0.20.0
